+++ packages/rosen-extractor/tests/getRosenData/bitcoin-runes/bitcoinRunesEsploraRosenExtractor.spec.ts
@@ -65,8 +65,8 @@ describe('BitcoinRunesEsploraRosenExtractor', () => {
     });
 
     /**
-     * @target `BitcoinRunesEsploraRosenExtractor.get` should return undefined when
-     * the utxos are unorganized
+     * @target `BitcoinRunesEsploraRosenExtractor.get` should extract rosenData from
+     * Runes lock tx successfully when the utxos are unorganized
      * @dependencies
      * @scenario
      * - mock a tx with incorrectly ordered utxos
@@ -75,7 +75,7 @@ describe('BitcoinRunesEsploraRosenExtractor', () => {
      * @expected
      * - it should return expected rosenData object
      */
-    it('should return undefined when the utxos are unorganized', () => {
+    it('should extract rosenData from Runes lock tx successfully when the utxos are unorganized', () => {
       const unorderedTx = testData.txs.unorderedTx;
 
       const extractor = new BitcoinRunesEsploraRosenExtractor(
@@ -84,7 +84,7 @@ describe('BitcoinRunesEsploraRosenExtractor', () => {
       );
       const result = extractor.get(unorderedTx as BitcoinEsploraTransaction);
 
-      expect(result).toBeUndefined();
+      expect(result).toStrictEqual(testData.rosenData);
     });
 
     /**