+++ packages/extended-typeorm/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rosen-bridge/extended-typeorm
 
+## 1.0.0
+
+### Major Changes
+
+- Update node version to 22.18
+
+### Minor Changes
+
+- Update typeorm version to 0.3.26
+
 ## 0.2.1
 
 ### Patch Changes