+++ packages/tx-pot/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rosen-bridge/tx-pot
 
+## 2.0.0
+
+### Major Changes
+
+- Update node version to 22.18
+
+### Minor Changes
+
+- Update typeorm version to 0.3.26
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-logger@3.0.0
+  - @rosen-bridge/extended-typeorm@1.0.0
+
 ## 1.1.0
 
 ### Minor Changes